+++ src/main/java/org/morphix/convert/strategy/ConversionStrategy.java
@@ -12,15 +12,15 @@
  */
 package org.morphix.convert.strategy;
 
-import static java.util.stream.Collectors.toMap;
-
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Predicate;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import org.morphix.lang.function.BinaryOperators;
 import org.morphix.lang.function.Predicates;
 import org.morphix.reflection.ExtendedField;
 import org.morphix.reflection.Fields;
@@ -55,7 +55,7 @@ public interface ConversionStrategy {
 	 * @return all the source field object pairs to convert to destination
 	 */
 	default <T> List<ExtendedField> findAll(final T source) {
-		return findFields(source, noFilter()).toList();
+		return findFields(source).toList();
 	}
 
 	/**
@@ -71,8 +71,8 @@ public interface ConversionStrategy {
 	static <T> Stream<ExtendedField> findFields(final T obj, final Predicate<? super ExtendedField> filter) {
 		// TODO: check duplicate fields (with the same name) in hierarchy (fields/getters)
 		Map<String, ExtendedField> nameToFieldMap =
-				Fields.getAllDeclaredInHierarchy(obj.getClass(), MemberPredicates.<Field>isNotStatic()).stream()
-						.collect(toMap(Field::getName, field -> ExtendedField.of(field, obj), (field1, field2) -> field1));
+				Fields.getAllDeclaredInHierarchy(obj.getClass(), MemberPredicates.isNotStatic()).stream()
+						.collect(Collectors.toMap(Field::getName, field -> ExtendedField.of(field, obj), BinaryOperators.first()));
 
 		List<Method> getterMethods = Methods.getAllDeclaredInHierarchy(obj.getClass(), MethodType.GETTER.getPredicate());
 		for (Method getterMethod : getterMethods) {
@@ -85,6 +85,19 @@ public interface ConversionStrategy {
 			}
 		}
 		return nameToFieldMap.values().stream().filter(filter);
+	}
+
+	/**
+	 * Returns a stream of fields.<br>
+	 * All static fields are filtered out by default.
+	 *
+	 * @param <T> source object type
+	 *
+	 * @param obj object on which find fields
+	 * @return stream of fields
+	 */
+	static <T> Stream<ExtendedField> findFields(final T obj) {
+		return findFields(obj, noFilter());
 	}
 
 	/**