+++ src/test/java/org/morphix/lang/function/BinaryOperatorsTest.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package org.morphix.lang.function;
+
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+import org.junit.jupiter.api.Test;
+
+/**
+ * Test class for {@link BinaryOperators}.
+ *
+ * @author Radu Sebastian LAZIN
+ */
+class BinaryOperatorsTest {
+
+	private static final String FIRST = "first";
+	private static final String SECOND = "second";
+
+	@Test
+	void shouldReturnTheFirstValue() {
+		String first = BinaryOperators.<String>first().apply(FIRST, SECOND);
+
+		assertThat(first, equalTo(FIRST));
+	}
+
+	@Test
+	void shouldReturnTheSecondValue() {
+		String first = BinaryOperators.<String>second().apply(FIRST, SECOND);
+
+		assertThat(first, equalTo(SECOND));
+	}
+}